added sessionID request`s verification; added sessionClose request at server closing; added error icon for problem contacts; some fixes;

+++ contents/code/serv.py
@@ -11,6 +11,7 @@ class ServerDaemon():
 		self.Parent = parent
 		self.Parent.serverState = self.serverState
 		self.commonSetOfSharedSource = commonSetOfSharedSource
+		self.currentSessionID = {}	## {remoteServerAddr : sessionID}
 		try :
 			exception = False
 			self._srv = ThreadServer(serveraddr, allow_none = True, \
@@ -27,10 +28,12 @@ class ServerDaemon():
 			self._srv.register_function(self.requestSharedSourceStruct, 'requestSharedSourceStruct')
 			self._srv.register_function(self.requestAvatar, 'requestAvatar')
 
-	def sessionID(self):
-		print self._srv.client_address, '--sessionID'
+	def sessionID(self, clientIP = ''):
+		#print self._srv.client_address, clientIP, '--sessionID'
+		if clientIP != self._srv.client_address[0] : return None
 		fileName = randomString(DIGITS_LENGTH)
 		_id = randomString(DIGITS_LENGTH)
+		self.currentSessionID[clientIP] = _id
 		list_ = [fileName, _id, self.serverState, self.Parent.previousState]
 		DataRendering().listToFile(list_, fileName)
 		with open(Path.tempStruct(fileName), "rb") as handle :
@@ -38,15 +41,22 @@ class ServerDaemon():
 
 	def sessionClose(self, sessionID = ''):
 		print self._srv.client_address, '--sessionClose'
-		pass
+		print sessionID, self.currentSessionID[self._srv.client_address[0]]
+		addr = self._srv.client_address[0]
+		if sessionID != self.currentSessionID[addr] : return None
+		if addr in self.currentSessionID : del self.currentSessionID[addr]
 
 	def python_clean(self, name, sessionID = ''):
-		print self._srv.client_address, '--python_clean'
+		#print self._srv.client_address, '--python_clean'
+		#print sessionID, self.currentSessionID[self._srv.client_address[0]]
+		if sessionID != self.currentSessionID[self._srv.client_address[0]] : return None
 		path_ = Path.tempStruct(name)
 		if os.path.isfile(path_) : os.remove(path_)
 
 	def python_file(self, id_, sessionID = ''):
-		print self._srv.client_address, '--python_file'
+		#print self._srv.client_address, '--python_file'
+		#print sessionID, self.currentSessionID[self._srv.client_address[0]]
+		if sessionID != self.currentSessionID[self._srv.client_address[0]] : return None
 		""" добавить обработчик ошибок соединения и существования файлов """
 		#print id_, type(id_), str(self.commonSetOfSharedSource[int(id_)]), '  serv'
 
@@ -55,13 +65,17 @@ class ServerDaemon():
 					return xmlrpclib.Binary(handle.read())
 
 	def requestSharedSourceStruct(self, name, sessionID = ''):
-		print self._srv.client_address, '--requestSharedSourceStruct'
+		#print self._srv.client_address, '--requestSharedSourceStruct'
+		#print sessionID, self.currentSessionID[self._srv.client_address[0]]
+		if sessionID != self.currentSessionID[self._srv.client_address[0]] : return None
 		#print Path.multiPath(Path.tempStruct, 'server', name), ' in requestSharedSourceStruct'
 		with open(Path.multiPath(Path.tempStruct, 'server', name), "rb") as handle:
 			return xmlrpclib.Binary(handle.read())
 
 	def requestAvatar(self, sessionID = ''):
-		print self._srv.client_address, '--requestAvatar'
+		#print self._srv.client_address, '--requestAvatar'
+		#print sessionID, self.currentSessionID[self._srv.client_address[0]]
+		if sessionID != self.currentSessionID[self._srv.client_address[0]] : return None
 		with open(unicode(self.Parent.avatarPath), "rb") as handle:
 			return xmlrpclib.Binary(handle.read())
 
@@ -69,6 +83,7 @@ class ServerDaemon():
 		self._srv.serve_forever()
 
 	def _shutdown(self, str_ = ''):
+		self.currentSessionID.clear()
 		self._srv.shutdown()
 		print ' server terminated'
 		self.Parent.serverDown.emit(str_)
